requirements.txt
svgpathtools>=1.6.0
matplotlib>=3.5.0
scikit-learn>=1.2.0
numpy>=1.23.0
imageio>=2.25.0
networkx==3.0
pyserial==3.5
click==8.1.3
requests==2.28.2